requirements.txt
numpy==1.19.4
aiostream==0.4.1
pyzmq==19.0.2
pyyaml==5.3.1
plyvel==1.3.0
loguru==0.5.3
docker==4.4.0
kubernetes==12.0.1
redis==3.5.3
fastapi==0.63.0
uvicorn==0.13.4 
pydantic-yaml==0.2.3 
envyaml==1.7.210310 
